fix: Refactor all services to use correct ILuceneIndexService interface
- FileIndexingService: Refactored to use IndexDocumentsAsync instead of direct IndexWriter
- BatchIndexingService: Updated to use proper interface methods (delete + index pattern)
- FileWatcherService: Now uses IFileIndexingService interface
- Fixed type consistency: BatchIndexingStats.TotalDocuments is now long (not int)
- All services now properly registered in DI container
- Project builds successfully with all core services working

+++ COA.CodeSearch.Next.McpServer/Services/IBatchIndexingService.cs
@@ -32,7 +32,7 @@ public class BatchIndexingStats
 {
     public int PendingDocuments { get; set; }
     public int TotalBatches { get; set; }
-    public int TotalDocuments { get; set; }
+    public long TotalDocuments { get; set; }
     public TimeSpan AverageBatchTime { get; set; }
     public DateTime LastCommit { get; set; }
 }
